Correction to modules loading order
MODPROBE_OPTIONS contains important '--dirname' parameter (it's recorded
in the SCRIPT which is under /usr/src by default), and kernel could be
in %buildroot, which is under /usr/src too. And with current approach of
ext4 rootfs we build it at any time (at vm-run or at filetrigger) but
fresh /usr/src is bind mounted from 9p.

Also, see comment in the code.

Signed-off-by: Vitaly Chikunov <vt@altlinux.org>