Don't resize tty if stdio is not tty
Previously only stdin is checked, now sttin and stdout.

It was thought we need to disable tty handling if run in batch build.
This change made it useful even in interactive session if we 'cat'
binary data (for example kdump from rootfs) from VM with redirection
to a file on host.

Also, don't clutter kernel command line with 'NOTTY' signal to VM, use
SCRIPT.

Signed-off-by: Vitaly Chikunov <vt@altlinux.org>