Experimental virtiofs support
virtiofsd requires root privileges or user namespace to change uids. It
will err when a different vm uid will try to create files for example.

 [... ERROR virtiofsd::passthrough::credentials] failed to change uid back to root: Operation not permitted (os error 1)

`--sandbox` requires root to chroot or to create mount, pid, and net
namespaces, that's why it's turned off.

Running under root uid fails with:

  [... ERROR virtiofsd] can't apply the child capabilities: failed to sync capabilities with the kernel

Thus `LD_PRELOAD` is unset to disable fakeroot.

virtiofsd will automatically exit when QEMU disconnects (forgetting to
remove .pid file.) But we can't be sure it was ever connected, so kill
is necessary precaution.

Currently, boot can be tested with:

  vm-run --virtiofs --append="rootfstype=virtiofs root=vfsd"

Invocation is logged in a similar way as for the qemu-system binary, so
the user knows how to reproduce the run manually.

Link: https://gitlab.com/virtio-fs/virtiofsd
Signed-off-by: Vitaly Chikunov <vt@altlinux.org>