Experimental: Remove -nographic if there is DISPLAY
`-nographic` interferes with DISPLAY. Perhaps, user wants graphical
output if they managed to pass DISPLAY into Hasher (with hsh-shell -Y).
If not, manually add `-nographic` or unset DISPLAY.

This should be transparent for anybody, except for experimenting. But,
maybe better is to add CLI option for this, or even to add more options,
like `-vga virtio -display gtk`, and change console= or add vga console
support.

With this the way to run -display gtk:

  hsh-install fonts-ttf-freefont fontconfig xauth
  hsh-install kernel-modules-drm-@flavour@
  hsh-install qemu-ui-gtk qemu-device-display-virtio-{vga,gpu}
  hsh-shell-kvm -Y
  vm-run --sandbox=off -vga virtio -display gtk --modules=virtio-gpu

Signed-off-by: Vitaly Chikunov <vt@altlinux.org>